requirements.txt
openai-whisper==20231114
pyttsx3==2.90
rich==13.5.2
pyfiglet==0.8.post0
python-dotenv==1.0.0
numpy
sounddevice
soundfile
